+++ src/main/java/cn/bilicraft/bccyberware/config/model/ResourcePackDeploymentType.java
@@ -1,7 +1,0 @@
-package cn.bilicraft.bccyberware.config.model;
-
-public enum ResourcePackDeploymentType {
-    RESOURCE_PACK_MANAGER,
-    SELFHOST,
-    EXTERNAL
-}